+++ translations/pl/controllers.php
@@ -26,5 +26,5 @@ return [
     'You cannot delete currently logged-in user.' => 'Nie możesz usunąć aktualnie zalogowanego użytkownika.',
     'Your email address has been confirmed!' => 'Twój adres e-mail został potwierdzony!',
     'Your password has been changed successfully.' => 'Twoje hasło zostało pomyślnie zmienione.',
-    'You’ve successfully logged out. For added security, consider closing all browser windows.' => 'Wylogowano pomyślnie. Dla większego bezpieczeństwa warto zamknąć wszystkie okna przeglądarki.',
+    'You’ve successfully logged out. Close all browser windows.' => 'Wylogowano Cię pomyślnie. Zamknij wszystkie okna przeglądarki.',
 ];